Share common test utils
