+++ src/tests/TestProviders.tsx
@@ -1,0 +1,46 @@
+import React from "react";
+import { RouterContext } from "next/dist/next-server/lib/router-context";
+import { NextRouter } from "next/router";
+import { MockedProvider, MockedResponse } from "@apollo/client/testing";
+
+const mockRouter: NextRouter = {
+  basePath: "",
+  pathname: "/",
+  route: "/",
+  asPath: "/",
+  query: {},
+  push: jest.fn(() => Promise.resolve(true)),
+  prefetch: jest.fn(() => Promise.resolve()),
+  replace: jest.fn(() => Promise.resolve(true)),
+  back: jest.fn(() => Promise.resolve(true)),
+  reload: jest.fn(() => Promise.resolve(true)),
+  beforePopState: jest.fn(() => Promise.resolve(true)),
+  events: {
+    on: jest.fn(),
+    off: jest.fn(),
+    emit: jest.fn(),
+  },
+  isFallback: false,
+  isLocaleDomain: true,
+  isPreview: false,
+  isReady: true,
+  locale: "fi",
+  defaultLocale: "fi",
+};
+
+type Props = {
+  mocks?: ReadonlyArray<MockedResponse>;
+  children: React.ReactNode;
+};
+
+function TestProviders({ mocks, children }: Props) {
+  return (
+    <RouterContext.Provider value={mockRouter}>
+      <MockedProvider mocks={mocks} addTypename={false}>
+        {children}
+      </MockedProvider>
+    </RouterContext.Provider>
+  );
+}
+
+export default TestProviders;
